# Test performance of first query