--- a/sqlitecluster/SQLite.cpp
+++ b/sqlitecluster/SQLite.cpp
@@ -150,10 +150,6 @@ sqlite3* SQLite::initializeDB(const string& filename, int64_t mmapSizeGB, bool h
         SERROR("sqlite3_open_v2 returned " << result << ", Extended error code: " << sqlite3_extended_errcode(db));
     }
 
-    // PRAGMA legacy_file_format=OFF sets the default for creating new databases, so it must be called before creating
-    // any tables to be effective.
-    SASSERT(!SQuery(db, "new file format for DESC indexes", "PRAGMA legacy_file_format = OFF"));
-
     return db;
 }
 
@@ -205,11 +201,11 @@ void SQLite::commonConstructorInitialization(bool hctree) {
 
     // WAL is what allows simultaneous read/writing.
     if (!hctree) {
-        SASSERT(!SQuery(_db, "enabling write ahead logging (wal2)", "PRAGMA journal_mode = wal2;"));
+        SASSERT(!_wrapSQuery(_db, "enabling write ahead logging (wal2)", "PRAGMA journal_mode = wal2;"));
     }
 
     if (_mmapSizeGB) {
-        SASSERT(!SQuery(_db, "enabling memory-mapped I/O", "PRAGMA mmap_size=" + to_string(_mmapSizeGB * 1024 * 1024 * 1024) + ";"));
+        SASSERT(!_wrapSQuery(_db, "enabling memory-mapped I/O", "PRAGMA mmap_size=" + to_string(_mmapSizeGB * 1024 * 1024 * 1024) + ";"));
     }
 
     // Enable tracing for performance analysis.
@@ -218,7 +214,7 @@ void SQLite::commonConstructorInitialization(bool hctree) {
     // Update the cache. -size means KB; +size means pages
     if (_cacheSize) {
         SINFO("Setting cache_size to " << _cacheSize << "KB");
-        SQuery(_db, "increasing cache size", "PRAGMA cache_size = -" + SQ(_cacheSize) + ";");
+        _wrapSQuery(_db, "increasing cache size", "PRAGMA cache_size = -" + SQ(_cacheSize) + ";");
     }
 
     // Register the authorizer callback which allows callers to whitelist particular data in the DB.
@@ -432,7 +428,7 @@ bool SQLite::beginTransaction(SQLite::TRANSACTION_TYPE type) {
 
     SINFO("Beginning transaction - open transaction count: " << (_sharedData.openTransactionCount));
     uint64_t before = STimeNow();
-    _insideTransaction = !SQuery(_db, "starting db transaction", "BEGIN CONCURRENT");
+    _insideTransaction = !_wrapSQuery(_db, "starting db transaction", "BEGIN CONCURRENT");
 
     // Because some other thread could commit once we've run `BEGIN CONCURRENT`, this value can be slightly behind
     // where we're actually able to start such that we know we shouldn't get a conflict if this commits successfully on
@@ -547,7 +543,7 @@ bool SQLite::read(const string& query, SQResult& result, bool skipInfoWarn) cons
         queryResult = true;
     } else {
         _isDeterministicQuery = true;
-        queryResult = !SQuery(_db, "read only query", query, result, 2000 * STIME_US_PER_MS, skipInfoWarn);
+        queryResult = !_wrapSQuery(_db, "read only query", query, result, 2000 * STIME_US_PER_MS, skipInfoWarn);
         if (_isDeterministicQuery && queryResult && insideTransaction()) {
             _queryCache.emplace(make_pair(query, result));
         }
@@ -637,7 +633,7 @@ bool SQLite::_writeIdempotent(const string& query, SQResult& result, bool always
 
     // First, check our current state
     SQResult results;
-    SASSERT(!SQuery(_db, "looking up schema version", "PRAGMA schema_version;", results));
+    SASSERT(!_wrapSQuery(_db, "looking up schema version", "PRAGMA schema_version;", results));
     SASSERT(!results.empty() && !results[0].empty());
     uint64_t schemaBefore = SToUInt64(results[0][0]);
     uint64_t changesBefore = sqlite3_total_changes(_db);
@@ -650,17 +646,17 @@ bool SQLite::_writeIdempotent(const string& query, SQResult& result, bool always
     {
         shared_lock<shared_mutex> lock(_sharedData.writeLock);
         if (_enableRewrite) {
-            resultCode = SQuery(_db, "read/write transaction", query, result, 2'000'000, true);
+            resultCode = _wrapSQuery(_db, "read/write transaction", query, result, 2'000'000, true);
             if (resultCode == SQLITE_AUTH) {
                 // Run re-written query.
                 _currentlyRunningRewritten = true;
                 SASSERT(SEndsWith(_rewrittenQuery, ";"));
-                resultCode = SQuery(_db, "read/write transaction", _rewrittenQuery);
+                resultCode = _wrapSQuery(_db, "read/write transaction", _rewrittenQuery);
                 usedRewrittenQuery = true;
                 _currentlyRunningRewritten = false;
             }
         } else {
-            resultCode = SQuery(_db, "read/write transaction", query, result);
+            resultCode = _wrapSQuery(_db, "read/write transaction", query, result);
         }
     }
 
@@ -678,7 +674,7 @@ bool SQLite::_writeIdempotent(const string& query, SQResult& result, bool always
     }
 
     // See if the query changed anything
-    SASSERT(!SQuery(_db, "looking up schema version", "PRAGMA schema_version;", results));
+    SASSERT(!_wrapSQuery(_db, "looking up schema version", "PRAGMA schema_version;", results));
     SASSERT(!results.empty() && !results[0].empty());
     uint64_t schemaAfter = SToUInt64(results[0][0]);
     uint64_t changesAfter = sqlite3_total_changes(_db);
@@ -701,7 +697,7 @@ bool SQLite::prepare(uint64_t* transactionID, string* transactionhash) {
 
     // Look up the oldest commit in our chosen journal, and compute the oldest commit we intend to keep.
     SQResult journalLookupResult;
-    SASSERT(!SQuery(_db, "getting commit min", "SELECT MIN(id) FROM " + _journalName, journalLookupResult));
+    SASSERT(!_wrapSQuery(_db, "getting commit min", "SELECT MIN(id) FROM " + _journalName, journalLookupResult));
     uint64_t minJournalEntry = journalLookupResult.size() ? SToUInt64(journalLookupResult[0][0]) : 0;
 
     // Note that this can change before we hold the lock on _sharedData.commitLock, but it doesn't matter yet, as we're only
@@ -719,7 +715,7 @@ bool SQLite::prepare(uint64_t* transactionID, string* transactionhash) {
     if (minJournalEntry < oldestCommitToKeep) {
         shared_lock<shared_mutex> lock(_sharedData.writeLock);
         string query = "DELETE FROM " + _journalName + " WHERE id < " + SQ(oldestCommitToKeep) + " LIMIT " + SQ(deleteLimit);
-        SASSERT(!SQuery(_db, "Deleting oldest journal rows", query));
+        SASSERT(!_wrapSQuery(_db, "Deleting oldest journal rows", query));
     }
 
     // We lock this here, so that we can guarantee the order in which commits show up in the database.
@@ -767,7 +763,7 @@ bool SQLite::prepare(uint64_t* transactionID, string* transactionhash) {
         SINFO("Will commmit blank query");
     }
 
-    int result = SQuery(_db, "updating journal", query);
+    int result = _wrapSQuery(_db, "updating journal", query);
     _prepareElapsed += STimeNow() - before;
     if (result) {
         // Couldn't insert into the journal; roll back the original commit
@@ -804,7 +800,7 @@ int SQLite::commit(const string& description, function<void()>* preCheckpointCal
     _conflictTable = "";
     uint64_t before = STimeNow();
     uint64_t beforeCommit = STimeNow();
-    result = SQuery(_db, "committing db transaction", "COMMIT");
+    result = _wrapSQuery(_db, "committing db transaction", "COMMIT");
     _lastConflictPage = _conflictPage;
     _lastConflictTable = _conflictTable;
 
@@ -907,7 +903,7 @@ void SQLite::rollback() {
                 SINFO("Rolling back transaction: " << _uncommittedQuery.substr(0, 100));
             }
             uint64_t before = STimeNow();
-            SASSERT(!SQuery(_db, "rolling back db transaction", "ROLLBACK"));
+            SASSERT(!_wrapSQuery(_db, "rolling back db transaction", "ROLLBACK"));
             _rollbackElapsed += STimeNow() - before;
         }
 
@@ -989,7 +985,7 @@ int SQLite::getCommits(uint64_t fromIndex, uint64_t toIndex, SQResult& result, u
     if (timeoutLimitUS) {
         setTimeout(timeoutLimitUS);
     }
-    int queryResult = SQuery(_db, "getting commits", query, result);
+    int queryResult = _wrapSQuery(_db, "getting commits", query, result);
     clearTimeout();
     return queryResult;
 }
@@ -1176,7 +1172,7 @@ void SQLite::setUpdateNoopMode(bool enabled) {
 
     // Enable or disable this query.
     string query = "PRAGMA noop_update="s + (enabled ? "ON" : "OFF") + ";";
-    SQuery(_db, "setting noop-update mode", query);
+    _wrapSQuery(_db, "setting noop-update mode", query);
     _noopUpdateMode = enabled;
 
     // If we're inside a transaction, make sure this gets saved so it can be replicated.
@@ -1229,7 +1225,7 @@ int SQLite::getPreparedStatements(const string& query, list<sqlite3_stmt*>& stat
 void SQLite::setQueryOnly(bool enabled) {
     SQResult result;
     string query = "PRAGMA query_only = "s + (enabled ? "true" : "false") + ";";
-    SQuery(_db, "set query_only", query, result);
+    _wrapSQuery(_db, "set query_only", query, result);
 }
 
 int64_t SQLite::getLastConflictPage() const {
@@ -1285,7 +1281,7 @@ map<uint64_t, tuple<string, string, uint64_t>> SQLite::SharedData::popCommittedT
     return result;
 }
 
-int SQLite::_wrapSQuery(sqlite3* db, const char* e, const string& sql, SQResult& result, int64_t warnThreshold, bool skipInfoWarn)
+int SQLite::_wrapSQuery(sqlite3* db, const char* e, const string& sql, SQResult& result, int64_t warnThreshold, bool skipInfoWarn) const
 {
     uint64_t startTime = 0;
     if (_isFirstQuery && !skipInfoWarn) {
@@ -1310,7 +1306,7 @@ int SQLite::_wrapSQuery(sqlite3* db, const char* e, const string& sql, SQResult&
     return squeryResult;
 }
 
-int SQLite::_wrapSQuery(sqlite3* db, const char* e, const string& sql, int64_t warnThreshold, bool skipInfoWarn)
+int SQLite::_wrapSQuery(sqlite3* db, const char* e, const string& sql, int64_t warnThreshold, bool skipInfoWarn) const
 {
     uint64_t startTime = 0;
     if (_isFirstQuery && !skipInfoWarn) {