### Details
This separates out slow query logging to log the first queries on each SQLite object separately.

### Fixed Issues
Diagnosing: https://github.com/Expensify/Expensify/issues/458560#issuecomment-2934250193

### Tests

You can just run the tests and watch the logs.

Currently though, all of our first queries seem to be a pragma:
```
2025-06-01T10:54:00.817342+00:00 expensidev-2404 bedrock10010: n0ZpGx (SQLite.cpp:1326) _wrapSQuery [socket30] [info] First query completed (12ms): PRAGMA journal_mode = wal2;
2025-06-01T10:54:00.817420+00:00 expensidev-2404 bedrock10010: rzpLFf (SQLite.cpp:1326) _wrapSQuery [socket31] [info] First query completed (11ms): PRAGMA journal_mode = wal2;
2025-06-01T10:54:03.182839+00:00 expensidev-2404 bedrock10097: VVLlcA (SQLite.cpp:1326) _wrapSQuery [socket20] [info] First query completed (10ms): PRAGMA journal_mode = wal2;
2025-06-01T10:54:06.277918+00:00 expensidev-2404 bedrock10052: nR3ome (SQLite.cpp:1326) _wrapSQuery [socket3] [info] First query completed (12ms): PRAGMA journal_mode = wal2;
2025-06-01T10:54:28.900195+00:00 expensidev-2404 bedrock10010: 1emOIl (SQLite.cpp:1326) _wrapSQuery [socket201] [info] First query completed (11ms): PRAGMA journal_mode = wal2;
2025-06-01T10:54:43.465939+00:00 expensidev-2404 bedrock10010: n3vTV8 (SQLite.cpp:1326) _wrapSQuery [socket70] [info] First query completed (14ms): PRAGMA journal_mode = wal2;
```

_________
**Internal Testing Reminder:** when changing bedrock, please compile auth against your new changes